Add memhashable config structs, add overlapped descriptor region support, misc. changes

+++ src/phantasm-hardware-interface/common/common_fwd.hh
@@ -21,7 +21,7 @@ namespace phi
 template <class T, bool GenCheckEnabled>
 struct linked_pool;
 
-template <class T, uint8_t N>
+template <class T, uint32_t N>
 struct flat_vector;
 
 struct page_allocator;
